% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorea.R
\name{colorea}
\alias{colorea}
\title{Create a palette color}
\usage{
colorea(colpal, a)
}
\arguments{
\item{colpal}{A vector}

\item{a}{A number}
}
\value{
A palette color of length \code{colpal} and alpha \code{a}.
}
\description{
Create a palette color
}
\examples{
colorea(1,0.5)

}
