% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fulldata_sh}
\alias{fulldata_sh}
\title{Output from SAM year 2015}
\format{A list with 65 objects. For exmaple: \describe{ 
  \item{pobspal}{observed catch-at-age proportion from longline fishery}
  \item{SB}{anual spawning biomass} ... }}
\source{
\url{http://www.ifop.cl/}
}
\usage{
fulldata_sh
}
\description{
A dataset containing the severals variables like biomass, catch-at-age 
observations, populations parameters, etc. derived from a stock assessment 
model carried out in Chile during 2014. Juan-Carlos Quiroz.
}
\keyword{datasets}

