% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hakeRead.R
\docType{package}
\name{hakeRead}
\alias{hakeRead}
\alias{hakeRead-package}
\title{hakeRead: A package for computating the notorious bar statistic.}
\description{
The hakeRead package provides three categories of important functions:
foo, bar and baz.
}
\section{hakeRead functions}{

The foo functions ...
}

