% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.admb.R
\name{read.admb}
\alias{read.admb}
\title{Read a set of ADMB outcomes}
\usage{
read.admb(ifile)
}
\arguments{
\item{ifile}{list of ADMB files after convergence}
}
\value{
list of ADMB objects
}
\description{
\code{read.admb} returns the sum of all the values present in its arguments.
}
\details{
This is a generic function: methods can be defined for it directly or via the
\code{\link{read.psv}} group generic. For this to work properly, the
arguments \code{...} should be unnamed, and dispatch is on the first
argument.
}
\seealso{
\code{\link{prod}} for products, \code{\link{cumsum}} for cumulative
  sums, and \code{\link{colSums}}/\code{\link{rowSums}} marginal sums over 
  high-dimensional arrays.

Other aggregate functions: \code{\link{read.fit}},
  \code{\link{read.psv}}, \code{\link{read.rep}}
}
