% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.rep.R
\name{read.rep}
\alias{read.rep}
\title{Read a ADMB report file}
\usage{
read.rep(fn)
}
\arguments{
\item{fn}{list of ADMB files after convergence}
}
\value{
list of ADMB fits
}
\description{
\code{read.rep} returns the sum of all the values present in its arguments.
}
\details{
This is a generic function: methods can be defined for it directly or via the
\code{\link{read.fit}} group generic. For this to work properly, the
arguments \code{...} should be unnamed, and dispatch is on the first
argument.
}
\seealso{
\code{\link{prod}} for products, \code{\link{cumsum}} for cumulative
  sums, and \code{\link{colSums}}/\code{\link{rowSums}} marginal sums over 
  high-dimensional arrays.

Other aggregate.functions: \code{\link{read.admb}},
  \code{\link{read.fit}}, \code{\link{read.psv}}
}

